PKG -- [fcl] move config to authn, remove client.extensions
