+++ packages/fcl/src/current-user/exec-service/index.js
@@ -1,4 +1,3 @@
-import {config as fclConfig} from "@onflow/config"
 import {execHttpPost} from "./strategies/http-post"
 import {execIframeRPC} from "./strategies/iframe-rpc"
 import {execPopRPC} from "./strategies/pop-rpc"
@@ -6,8 +5,6 @@ import {execTabRPC} from "./strategies/tab-rpc"
 import {execExtRPC} from "./strategies/ext-rpc"
 import {execWcRPC} from "./strategies/wc-rpc"
 import {invariant} from "@onflow/util-invariant"
-import {configLens} from "../../config-utils"
-import {VERSION} from "../../VERSION"
 
 const STRATEGIES = {
   "HTTP/RPC": execHttpPost,
@@ -19,39 +16,14 @@ const STRATEGIES = {
   "WC/RPC": execWcRPC,
 }
 
-const makeDiscoveryServices = servicePlugin => {
-  // get services from any service plugins
-  const discoveryServices = servicePlugin ? servicePlugin.services : null
-  return discoveryServices
-}
-
 export async function execService({service, msg = {}, config = {}, opts = {}}) {
   msg.data = service.data
-  let discoveryServices = []
-  const {servicePlugin} = await fclConfig.get("wc.adapter", {
-    servicePlugin: null,
-  })
-  if (service.type === "authn") {
-    discoveryServices = makeDiscoveryServices(servicePlugin)
-  }
-  const fullConfig = {
-    ...config,
-    services: await configLens(/^service\./),
-    app: await configLens(/^app\.detail\./),
-    client: {
-      fclVersion: VERSION,
-      fclLibrary: "https://github.com/onflow/fcl-js",
-      hostname: window?.location?.hostname ?? null,
-      extensions: window?.fcl_extensions || [],
-      discoveryServices,
-    },
-  }
 
   try {
     const res = await STRATEGIES[service.method]({
       service,
       body: msg,
-      config: fullConfig,
+      config,
       opts,
     })
     if (res.status === "REDIRECT") {
@@ -63,7 +35,7 @@ export async function execService({service, msg = {}, config = {}, opts = {}}) {
         service: res.data,
         msg,
         opts,
-        config: fullConfig,
+        config,
       })
     } else {
       return res
